refactor: move selectors into a folder, add containerSelectors
